requirements.txt
openai-whisper>=20230918
scikit-learn>=1.3
sentence-transformers>=2.2
camel-tools>=1.5
click>=8.1
matplotlib>=3.7
seaborn>=0.12
numpy>=1.24
fastapi>=0.104
uvicorn[standard]>=0.24
python-multipart>=0.0.6
